Combined playlist building into TimerTrigger
