+++ TimerTrigger1/__init__.py
@@ -3,6 +3,8 @@ import logging
 
 import azure.functions as func
 
+import playlistBuilder as build
+
 
 def main(mytimer: func.TimerRequest) -> None:
     utc_timestamp = datetime.datetime.utcnow().replace(
@@ -12,3 +14,5 @@ def main(mytimer: func.TimerRequest) -> None:
         logging.info('The timer is past due!')
 
     logging.info('Python timer trigger function ran at %s', utc_timestamp)
+
+    build.mainBuild()
